Remove test files; Add new files for displaying the main app page; Modify login.js so that it calls functions based on user/admin role
