+++ frontend/index.html
@@ -23,11 +23,7 @@
       <button>Register</button>
     </div>
 
-    <section id="movies-container">
-      <h2>Movies</h2>
-      <!-- ode se prikazuju filmovi pomocu JS -->
-      <button id="logout-btn">Logout</button>
-    </section>
+    <section id="movies-container"></section>
     <script type="module" src="./scripts/modules/index.js"></script>
   </body>
 </html>
