+++ frontend/ivona-test.html
@@ -18,6 +18,6 @@
         <!-- ode se prikazuju filmovi pomocu JS -->
       </section>
     </main>
-    <script type="module" src="./ivona-test/ivona-api.js"></script>
+    <script type="module" src="./scripts/modules/user-app.js"></script>
   </body>
 </html>
