+++ frontend/scripts/modules/index.js
@@ -1,3 +1,2 @@
-
 import "./login.js";
 import "./register.js";
