+++ frontend/scripts/modules/login.js
@@ -1,16 +1,19 @@
-
 import { generateJWT, saveAuthToken } from "./auth.js";
 import { mockUsers } from "../../marul-test/marul-mock-data.js";
+import { initUserApp } from "./user-app.js";
 
 document.addEventListener("DOMContentLoaded", function () {
   const loginContainer = document.getElementById("login-container");
-  if (!loginContainer) return; 
+  if (!loginContainer) return;
 
-  
   const loginButton = loginContainer.querySelector("button");
-  loginButton.addEventListener("click", function () {
-    const usernameInput = loginContainer.querySelector("input[placeholder='Username']");
-    const passwordInput = loginContainer.querySelector("input[placeholder='Password']");
+  loginButton.addEventListener("click", async function () {
+    const usernameInput = loginContainer.querySelector(
+      "input[placeholder='Username']"
+    );
+    const passwordInput = loginContainer.querySelector(
+      "input[placeholder='Password']"
+    );
     const username = usernameInput.value.trim();
     const password = passwordInput.value.trim();
 
@@ -19,16 +22,30 @@ document.addEventListener("DOMContentLoaded", function () {
       return;
     }
 
-    let registeredUsers = JSON.parse(localStorage.getItem("registeredUsers")) || []; 
+    let registeredUsers =
+      JSON.parse(localStorage.getItem("registeredUsers")) || [];
     const allUsers = [...mockUsers, ...registeredUsers];
 
     //test sa mock podacima
-    const user = allUsers.find(u => u.username === username && u.password === password);
+    const user = allUsers.find(
+      (u) => u.username === username && u.password === password
+    );
     if (user) {
       const token = generateJWT(user);
       saveAuthToken(token);
-      //preusmjeravanje na landing page
-      window.location.href = "ivona-test.html";
+
+      //! ode treba pozvat ralizite funkcije ovisno o tome je li admin ili ne
+      loginContainer.style.display = "none";
+      try {
+        if (user.admin) {
+          const adminModule = await import("./admin-app.js");
+          adminModule.initAdminApp();
+        } else {
+          initUserApp();
+        }
+      } catch (error) {
+        console.error("Greška pri učitavanju modula:", error);
+      }
     } else {
       alert("Neispravno korisničko ime ili lozinka.");
     }
