+++ frontend/scripts/modules/movie-details.js
